Reorg physical optimizer (incl. local)
